requirements.txt
Flask
numpy
pandas
matplotlib
mysql-connector-python
python-dotenv
gunicorn
psycopg2-binary
